requirement.txt
baidu-aip==1.3.1.1
browsermob-proxy==0.8.0
chardet==2.3.0
olefile==0.44
Pillow==4.0.0
PyMouse==1.0
pyocr==0.4.6
requests==2.13.0
selenium==3.3.1
six==1.10.0
